Feature: Progressive Messages (#591)
This change enables progressive rendering of messages as they are returned from the service. It also changes the state changes for a query such that the a query will always get a QueryComplete notification if the query is started successfully, and will always get an Error if the query fails to start. Batch start notifications are changed to be sent to the preview pane as a message with a link property. Correspondingly, the handler for changing the text selection in the query now uses a GET handler instead of a POST handler, with the URI being generated in the SQL output provider.