Fix/html encode display values (#604)
Adding a fairly robust utility function to encode results from databases with HTML entities -- since JS doesn't have a built in one. This fixes issues where displaying result sets with HTML would fail. It also fixes an undiscovered issue where tooltips for cells would be cut off if they contained a `'` or `"`. Since the change is in the presentation layer, copy/paste and save as functionality still works as expected.