Support Copy&Paste of Messages (#656)
Fixes #385. Added keyboard shortcut Ctrl+C and context menu support for copying from the Messages section. To do this:

Messages have a context menu handler with a single option of Copy.
Added use of the Rangy library which simplifies selection and text range manipulation, including pretty print support for copying as text. We need to copy as text since the HTML includes our anchor links to the current page and in dark mode the text is copied as black text on a black background

- Copy works with validation that only messages are copied
- Using Rangy's text() function to get correctly formatted text. This ensures it is tab delimited which is copiable into Excel and text editors in the expected format
- Copy shortcut key supports copying Messages